Add a chat feature. Step 1
A chat feature would allow any user to open a chat box with another user. Here we add a pop-up chat window (chat-list component) with list of users to chat. This chat window is routed as secondary router that user could chat and use application simultaneously. In step 1 it is still not possible to start chating.